debug: clean up slightly
make assert-db parameter so can be saved and restored just like
other things e.g. term-cache.

Also add bug-assert which is like bug-on but is like Rosette
assert (asserts condition to be true instead of false).

bug-on is implemented in terms of bug-assert now.

Goal is to make debugging support for assertions nice enough that
specs don't have to do odd things like add "check-unsat?" or "verify"
everywhere to their code in order to get nice error messages.