Tweak the way gpr-ref and gpr-set! work.
Right now we use a function to convert from register name to integer
index, use that to index into a vector.

This can unnecessarily drag theory of integers into the resulting
SMT query even though there's almost no reason to treat register
names as integers (e.g., no ordering, no arithmetic on them, etc.)

This changes general purpose registers to be a struct just like
CSRs, and dispatches to the correct register according to the name
used.

We keep around gpr->idx to be able to encode RISC-V instructions
as bytes, which is the only time the numeric value is actually
relevant.

The top-level bug that caused me to fix this was that we were unable
to use Boolector to synthesize RISC-V programs because it was being
(unncessarily) forced to reason about integers, which Boolector
cannot do.