Fix Makefrag bug and add slti/slt
Comments in Makefiles are continued to the next line when the line
ends in a backslash. I didn't catch this before until I opened the
file in an editor with highlighting for comments :(

Fixing the problem, I also noticed we didn't support slt/slti
so add those instructions as well.