add memmgr.rkt to lib.
This is the first step in having reusable, configurable memory
models in Serval. For now, only factor out the model that was being
used by the RISC-V interpreter (typed, disjoint regions of memory
where pointers are bitvectors).

Right now the code is a bit hacky, but it appears to work.
It's based on code I stole from @xi's memmgr code.

It defines a new generic type, memmgr, which can perform alloc,
load, store, and memset operations. There is only one such memmgr
for now, which implements the functionality that used to be implemented
in the RISC-V interpreter. (The interface is thus somewhat tailored
to RISC-V's use case, and we likely have to figure out something
more minimal, e.g., load and store should not take an address and
an offset).

init-cpu uses this is as the default memory manager, and I introduced
a backward-compatible, deprecated, cpu-mregions to make sure old
code still runs.

The memory management code is still quite messy---the implementation
is split across core.rkt and memmgr.rkt now with mblocks, mregions,
symbols, and globals all interacting in tightly-coupled ways.  Should
think about ways to factor this out even further.