riscv: all atomic instructions + add memory tests
We needed amoadd.{d,w} in order to say anything about BPF_XADD, and I
added the rest because why not.

Some memory tests (sh, sw, l*) cannot run because they use initialized
data in the code, and we don't have an ELF parser.

None of the atomic tests require it though so all of those are enabled.