serval/bpf: Redo the BPF memory model.
This removes the notions of pointers from the BPF interpreter.
Now pointers are just bitvectors (like riscv,x32) and the interpreter
delegates to a memory manager for memory instructions.

This removes some functionality from the BPF interpreter; it no longer
detects runtime pointer bugs since this should be the job of the memory
manager anyways.

BPF for now uses a flat memory model which is good enough for passing
the tests.