memmgr: add memmgr-atomic-begin and memmgr-atomic-end
This gives a way for memory managers to know which memory operations
are atomic. For most uses, this is irrelevant so we provide a default
implementation of (void). For BPF JIT verification, we can append
atomic begin and end entries to a trace to make sure that an atomic BPF
instruction is compiled to an atomic target instruction.