bpf: truncate address to underlying memory model bitwidth
This is due to a bit of an oddity in the BPF specification that the
memory model (including bitwidth, etc.) is purely determined by the
model of the underlying target architecture.

It should be the job of the bpf interpreter to do this truncation instead
of the memory model itself.