llvm: add very basic inttoptr support
Use the same mechanism as used in the RISC-V interpreter for converting
integers to mregions via guesses (which are then validated on access).

For stores, always convert pointers to integers first; for loads, convert
back to pointer using inttoptr if the type of the load is a pointer.

It's not super well tested, will have to write some additional tests
to make sure this is working right.